File structure refactor. Initial tests for sending entire led strip byte array
Preparing project to do all the processing on the host, and sending just the byte array of the entire LED strip status. What this will allow us to do in the the future is create a GUI to toggle LED's individually from the host, and will allow for a flexible way to send patterns through the LED strip without writing code on the hardware that controls the LED strip.